+++ src/daemon/classes/fs_data.cpp
@@ -221,4 +221,69 @@ FsData::parallax_size_md(unsigned int size_md) {
             size_md * 1024ull * 1024ull * 1024ull);
 }
 
+const std::shared_ptr<gkfs::utils::Stats>&
+FsData::stats() const {
+    return stats_;
+}
+
+void
+FsData::stats(const std::shared_ptr<gkfs::utils::Stats>& stats) {
+    FsData::stats_ = stats;
+}
+
+void
+FsData::close_stats() {
+    stats_.reset();
+}
+
+bool
+FsData::enable_stats() const {
+    return enable_stats_;
+}
+
+void
+FsData::enable_stats(bool enable_stats) {
+    FsData::enable_stats_ = enable_stats;
+}
+
+bool
+FsData::enable_chunkstats() const {
+    return enable_chunkstats_;
+}
+
+void
+FsData::enable_chunkstats(bool enable_chunkstats) {
+    FsData::enable_chunkstats_ = enable_chunkstats;
+}
+
+bool
+FsData::enable_prometheus() const {
+    return enable_prometheus_;
+}
+
+void
+FsData::enable_prometheus(bool enable_prometheus) {
+    FsData::enable_prometheus_ = enable_prometheus;
+}
+
+const std::string&
+FsData::stats_file() const {
+    return stats_file_;
+}
+
+void
+FsData::stats_file(const std::string& stats_file) {
+    FsData::stats_file_ = stats_file;
+}
+
+const std::string&
+FsData::prometheus_gateway() const {
+    return prometheus_gateway_;
+}
+
+void
+FsData::prometheus_gateway(const std::string& prometheus_gateway) {
+    FsData::prometheus_gateway_ = prometheus_gateway;
+}
+
 } // namespace gkfs::daemon