+++ src/daemon/handler/srv_metadata.cpp
@@ -41,6 +41,7 @@
 #include <daemon/ops/metadentry.hpp>
 
 #include <common/rpc/rpc_types.hpp>
+#include <common/statistics/stats.hpp>
 
 using namespace std;
 
@@ -84,6 +85,7 @@ rpc_srv_create(hg_handle_t handle) {
                                       __func__, e.what());
         out.err = -1;
     }
+
     GKFS_DATA->spdlogger()->debug("{}() Sending output err '{}'", __func__,
                                   out.err);
     auto hret = margo_respond(handle, &out);
@@ -94,6 +96,10 @@ rpc_srv_create(hg_handle_t handle) {
     // Destroy handle when finished
     margo_free_input(handle, &in);
     margo_destroy(handle);
+    if(GKFS_DATA->enable_stats()) {
+        GKFS_DATA->stats()->add_value_iops(
+                gkfs::utils::Stats::IopsOp::iops_create);
+    }
     return HG_SUCCESS;
 }
 
@@ -148,6 +154,11 @@ rpc_srv_stat(hg_handle_t handle) {
     // Destroy handle when finished
     margo_free_input(handle, &in);
     margo_destroy(handle);
+
+    if(GKFS_DATA->enable_stats()) {
+        GKFS_DATA->stats()->add_value_iops(
+                gkfs::utils::Stats::IopsOp::iops_stats);
+    }
     return HG_SUCCESS;
 }
 
@@ -241,6 +252,7 @@ rpc_srv_remove_metadata(hg_handle_t handle) {
             if(S_ISREG(md.mode()) && (md.size() != 0))
                 GKFS_DATA->storage()->destroy_chunk_space(in.path);
         }
+
     } catch(const gkfs::metadata::DBException& e) {
         GKFS_DATA->spdlogger()->error("{}(): path '{}' message '{}'", __func__,
                                       in.path, e.what());
@@ -265,6 +277,10 @@ rpc_srv_remove_metadata(hg_handle_t handle) {
     // Destroy handle when finished
     margo_free_input(handle, &in);
     margo_destroy(handle);
+    if(GKFS_DATA->enable_stats()) {
+        GKFS_DATA->stats()->add_value_iops(
+                gkfs::utils::Stats::IopsOp::iops_remove);
+    }
     return HG_SUCCESS;
 }
 
@@ -635,6 +651,10 @@ rpc_srv_get_dirents(hg_handle_t handle) {
     GKFS_DATA->spdlogger()->debug(
             "{}() Sending output response err '{}' dirents_size '{}'. DONE",
             __func__, out.err, out.dirents_size);
+    if(GKFS_DATA->enable_stats()) {
+        GKFS_DATA->stats()->add_value_iops(
+                gkfs::utils::Stats::IopsOp::iops_dirent);
+    }
     return gkfs::rpc::cleanup_respond(&handle, &in, &out, &bulk_handle);
 }